Revert "replace int for BigDecimal"
This reverts commit 5753571a90a2d2bbabb6ca0f0511a415e7540c06.

+++ src/main/java/lu/crx/financing/entities/Creditor.java
@@ -1,8 +1,6 @@
 package lu.crx.financing.entities;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
-
 import jakarta.persistence.Basic;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -39,6 +37,6 @@ public class Creditor implements Serializable {
      * Maximum acceptable financing rate for this creditor.
      */
     @Basic(optional = false)
-    private BigDecimal maxFinancingRateInBps;
+    private int maxFinancingRateInBps;
 
 }
