+++ src/main/java/lu/crx/financing/entities/PurchaserFinancingResult.java
@@ -3,14 +3,13 @@ package lu.crx.financing.entities;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
 @AllArgsConstructor
 public class PurchaserFinancingResult {
     private final Purchaser purchaser;
-    private final BigDecimal financingRateInBps;
+    private final int financingRateInBps;
     private final int financingTermInDays;
     private final LocalDate financingDate;
 }
